Support complex toolchains
Instead of specifying the name of the toolchain and relying on arm-eabi-
use the entire path to the custom toolchain.

Add the following line to BoardConfig.mk to use the custom toolchain for kernel
For linaro:
TARGET_KERNEL_CUSTOM_TOOLCHAIN := linaro/bin/arm-linux-gnueabihf-

For GCC 4.4.3:
TARGET_KERNEL_CUSTOM_TOOLCHAIN := arm-eabi-4.4.3/bin/arm-eabi-

The more we write out paths the better. This helped fix 4.6 compile of CM kernel for me
Change-Id: I45ed019278a1f310d6abc4c7fbfa14ddc6d184b7

Conflicts:

	core/tasks/kernel.mk